fixed: volume plot display error

+++ common/data_reader.py
@@ -17,5 +17,5 @@ def fread(f_name, spx, spy, spt):
         # big-endian: '>', signed: 'i', 16bit: '2'
         raw_data = np.fromfile(f, dtype='>i2').reshape((spx, spy, spt))
         # convert to regular array with best compatibility
-        raw_data = np.array(raw_data, dtype=np.float32)
+        raw_data = np.array(raw_data, dtype=np.int16)
         return raw_data
